+++ BACKEND/demo/src/main/java/com/example/demo/repository/UserRepository.java
@@ -6,8 +6,6 @@ import com.example.demo.entity.User;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long>{
+public interface UserRepository extends JpaRepository<User,Integer>{
     Optional<User> findByUsername(String username);
-
-
 }
