+++ BACKEND/demo/src/main/java/com/example/demo/service/UserService.java
@@ -30,7 +30,7 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public void deleteUser(Long userId) {
+    public void deleteUser(Integer userId) {
         boolean exists = userRepository.existsById(userId);
         if(!exists){
             throw new IllegalStateException("user with id" + userId+ "does not exists");
